Added animations

+++ ToDoList/Views/ListView.swift
@@ -11,19 +11,25 @@ struct ListView: View {
     @EnvironmentObject var listViewModel : ListViewModel
       
     var body: some View {
-        List {
-            ForEach(listViewModel.items){ item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item1: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemView()
+            } else {
+                List {
+                    ForEach(listViewModel.items){ item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item1: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(.plain)
         .navigationTitle("Todo list 📝")
         .navigationBarItems(leading: EditButton(), trailing: NavigationLink("Add", destination: {
             AddView()
